Refactor cubit to sry layout
