+++ .pre-commit-config.yaml
@@ -48,7 +48,7 @@ repos:
     hooks:
       - id: docformatter
         name: Run docformatter (formatter for docstrings)
-        args: ["--in-place", "--config=./pyproject.toml", "-r", "cubitpy/", "tests/", "tutorial/"]
+        args: ["--in-place", "--config=./pyproject.toml", "-r", "src/", "tests/"]
   - repo: https://github.com/econchick/interrogate
     rev: 1.7.0
     hooks:
