+++ src/cubitpy/cubit_utility.py
@@ -19,9 +19,31 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-# -----------------------------------------------------------------------------
-# -----------------------------------------------------------------------------
-"""This module contains extensions and utility functions for the cubit
-tutorial."""
+"""Utility functions for the use of cubitpy."""
 
-from .tutorial import cubit_step_by_step_tutorial_cli
+from cubitpy.conf import cupy
+
+
+def get_surface_center(surf):
+    """Get a 3D point that has the local coordinated on the surface of (0,0),
+    with the parameter space being ([-1,1],[-1,1])."""
+
+    if not surf.get_geometry_type() == cupy.geometry.surface:
+        raise TypeError("Did not expect {}".format(type(surf)))
+
+    range_u = surf.get_param_range_U()
+    u = 0.5 * (range_u[1] + range_u[0])
+    range_v = surf.get_param_range_V()
+    v = 0.5 * (range_v[1] + range_v[0])
+    return surf.position_from_u_v(u, v)
+
+
+def import_fluent_geometry(cubit, file, feature_angle=135):
+    """Import fluent mesh geometry in cubit from file with according
+    feature_angle."""
+
+    cubit.cmd(
+        'import fluent mesh geometry  "{}" feature_angle {} '.format(
+            file, feature_angle
+        )
+    )
